Fix backtick commands for execline 2.8.0.0
I'm the Alpine package maintainer for s6-overlay. Alpine's Edge
"bleeding edge" has execline 2.8.0.0 packaged and as a result
my s6-overlay 2.2.0.3 package is giving errors due to a change in
"backtick" behaviour as of execline 2.8.0.0 (detailed here:
http://skarnet.org/software/execline/upgrade.html).

This PR fixes the init-stage1 and init-stage2 scripts to work with the
newer version of execline.